refactor: rename autodiscover.py to autodiscovery.py to avoid conflict with function name
