+++ src/django_components/apps.py
@@ -10,7 +10,7 @@ class ComponentsConfig(AppConfig):
     # to Django's INSTALLED_APPS
     def ready(self) -> None:
         from django_components.app_settings import app_settings
-        from django_components.autodiscover import autodiscover, import_libraries
+        from django_components.autodiscovery import autodiscover, import_libraries
         from django_components.component_registry import registry
         from django_components.components.dynamic import DynamicComponent
         from django_components.template_loader import get_component_dirs
