+++ src/django_components/autodiscovery.py
